Split cancel and warning controllers
